Fix #5577: Solana/Filecoin network switch (#5681)
* Support fetching all networks for Ethereum, Solana and Filecoin coin types. Display all networks in the network picker (Solana networks first).

* Show alert when user switches to a coin type they do not have any accounts for. Reset back to previous chain if they choose to not create an account, otherwise show `AddAccountView` modal.

* Update known networks list until #5489 work is completed

* Set selected coin on wallet service when network changes

* Fix for selecting newly created account when a new keyring is created (ex. Solana keyring created), observe selected account changed to update network if needed.

+++ BraveWallet/Crypto/Stores/NetworkStore.swift
@@ -12,54 +12,76 @@ import Strings
 ///
 /// This wraps a JsonRpcService that you would obtain through BraveCore and makes it observable
 public class NetworkStore: ObservableObject {
-  @Published private(set) var ethereumChains: [BraveWallet.NetworkInfo] = []
+  
+  enum SetSelectedChainError: Error {
+    case selectedChainHasNoAccounts
+    case chainAlreadySelected
+    case unknown
+  }
+  
+  @Published private(set) var allChains: [BraveWallet.NetworkInfo] = []
+
   @Published private(set) var selectedChainId: String = BraveWallet.MainnetChainId
-
   var selectedChain: BraveWallet.NetworkInfo {
-    ethereumChains.first(where: { $0.chainId == self.selectedChainId }) ?? .init()
+    allChains.first(where: { $0.chainId == self.selectedChainId }) ?? .init()
   }
 
-  var selectedChainBinding: Binding<BraveWallet.NetworkInfo> {
-    .init(
-      get: { self.ethereumChains.first(where: { $0.chainId == self.selectedChainId }) ?? .init() },
-      set: {
-        self.setSelectedChain(chainId: $0.chainId, coin: $0.coin)
-      }
-    )
-  }
-
+  private let keyringService: BraveWalletKeyringService
   private let rpcService: BraveWalletJsonRpcService
   private let walletService: BraveWalletBraveWalletService
 
   public init(
+    keyringService: BraveWalletKeyringService,
     rpcService: BraveWalletJsonRpcService,
     walletService: BraveWalletBraveWalletService
   ) {
+    self.keyringService = keyringService
     self.rpcService = rpcService
     self.walletService = walletService
     self.updateChainList()
-    walletService.selectedCoin { [self] coin in
-      rpcService.chainId(coin) { chainId in
-        self.selectedChainId = chainId.isEmpty ? BraveWallet.MainnetChainId : chainId
-      }
-    }
     rpcService.add(self)
+    keyringService.add(self)
+    
+    Task { @MainActor in // fetch current selected network
+      let selectedCoin = await walletService.selectedCoin()
+      let chain = await rpcService.network(selectedCoin)
+      await setSelectedChain(chain)
+    }
   }
 
   private func updateChainList() {
-    walletService.selectedCoin { [self] coin in
-      rpcService.allNetworks(coin) { chains in
-        self.ethereumChains = chains.filter {
-          $0.chainId != BraveWallet.LocalhostChainId
+    Task { @MainActor in // fetch all networks for all coin types
+      self.allChains = await withTaskGroup(of: [BraveWallet.NetworkInfo].self) { [weak rpcService] group -> [BraveWallet.NetworkInfo] in
+        guard let rpcService = rpcService else { return [] }
+        for coinType in WalletConstants.supportedCoinTypes {
+          group.addTask {
+            let chains = await rpcService.allNetworks(coinType)
+            return chains.filter { // localhost not supported
+              $0.chainId != BraveWallet.LocalhostChainId
+            }
+          }
+        }
+        let allChains = await group.reduce([BraveWallet.NetworkInfo](), { $0 + $1 })
+        return allChains.sorted { lhs, rhs in
+          // sort solana chains to the front of the list
+          lhs.coin == .sol && rhs.coin != .sol
         }
       }
     }
   }
-
-  private func setSelectedChain(chainId: String, coin: BraveWallet.CoinType) {
-    guard self.selectedChainId != chainId else { return }
-    self.selectedChainId = chainId
-    self.rpcService.setNetwork(chainId, coin: coin) { _ in }
+  
+  @MainActor @discardableResult func setSelectedChain(_ network: BraveWallet.NetworkInfo) async -> SetSelectedChainError? {
+    let keyringId = network.coin.keyringId
+    let keyringInfo = await keyringService.keyringInfo(keyringId)
+    if keyringInfo.accountInfos.isEmpty {
+      // Need to prompt user to create new account via alert
+      return .selectedChainHasNoAccounts
+    } else {
+      guard self.selectedChainId != network.chainId else { return .chainAlreadySelected }
+      self.selectedChainId = network.chainId
+      let success = await rpcService.setNetwork(network.chainId, coin: network.coin)
+      return success ? nil : .unknown
+    }
   }
 
   // MARK: - Custom Networks
@@ -80,7 +102,7 @@ public class NetworkStore: ObservableObject {
         } else {
           // meaning add custom network failed for some reason.
           // Also add the the old network back on rpc service
-          if let oldNetwork = ethereumChains.first(where: { $0.id.lowercased() == network.id.lowercased() }) {
+          if let oldNetwork = allChains.filter({ $0.coin == .eth }).first(where: { $0.id.lowercased() == network.id.lowercased() }) {
             rpcService.addEthereumChain(oldNetwork) { _, _, _ in
               // Update `ethereumChains` by api calling
               self.updateChainList()
@@ -96,7 +118,7 @@ public class NetworkStore: ObservableObject {
     }
 
     isAddingNewNetwork = true
-    if ethereumChains.contains(where: { $0.id.lowercased() == network.id.lowercased() }) {
+    if allChains.filter({ $0.coin == .eth }).contains(where: { $0.id.lowercased() == network.id.lowercased() }) {
       removeNetworkForNewAddition(network) { [self] success in
         guard success else {
           isAddingNewNetwork = false
@@ -124,6 +146,10 @@ public class NetworkStore: ObservableObject {
     _ network: BraveWallet.NetworkInfo,
     completion: @escaping (_ success: Bool) -> Void
   ) {
+    guard network.coin == .eth else {
+      completion(false)
+      return
+    }
     rpcService.removeEthereumChain(network.id) { [self] success in
       if success {
         // check if its the current network, set mainnet the active net
@@ -145,6 +171,50 @@ extension NetworkStore: BraveWalletJsonRpcServiceObserver {
     updateChainList()
   }
   public func chainChangedEvent(_ chainId: String, coin: BraveWallet.CoinType) {
-    setSelectedChain(chainId: chainId, coin: coin)
+    walletService.setSelectedCoin(coin)
+    Task { @MainActor in
+      guard let chain = allChains.first(where: { $0.chainId == chainId && $0.coin == coin }) else {  return }
+      await setSelectedChain(chain)
+    }
+  }
+}
+
+extension NetworkStore: BraveWalletKeyringServiceObserver {
+  public func selectedAccountChanged(_ coin: BraveWallet.CoinType) {
+    Task { @MainActor in
+      // coin type of selected account might have changed
+      let chain = await rpcService.network(coin)
+      await setSelectedChain(chain)
+    }
+  }
+  
+  public func keyringCreated(_ keyringId: String) {
+    Task { @MainActor in
+      // select the newly created account
+      let keyring = await keyringService.keyringInfo(keyringId)
+      guard let newAccount = keyring.accountInfos.first else { return }
+      await keyringService.setSelectedAccount(newAccount.address, coin: newAccount.coin)
+    }
+  }
+  
+  public func keyringRestored(_ keyringId: String) {
+  }
+  
+  public func keyringReset() {
+  }
+  
+  public func locked() {
+  }
+  
+  public func unlocked() {
+  }
+  
+  public func backedUp() {
+  }
+  
+  public func accountsChanged() {
+  }
+  
+  public func autoLockMinutesChanged() {
   }
 }
